requirements.txt
--extra-index-url=https://test.pypi.org/simple/
aniso8601==8.0.0
atomicwrites==1.4.0
attrs==20.2.0
click==7.1.2
colorama==0.4.4
flasgger==0.9.5
Flask==1.1.2
flask-peewee==3.0.3
Flask-RESTful==0.3.8
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
MonacoRacing-ChyhrynYaroslav==0.0.6
packaging==20.4
peewee==3.14.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.1
python-simplexml==0.1.5
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wtf-peewee==3.0.2
WTForms==2.3.3
zipp==3.4.0
